Prevent AI-Induced Version Downgrades (Knowledge Cutoff Regression)
This change implements a deterministic validation layer and context-injection step to prevent AI agents from downgrading dependencies or modifying core runtime versions.

Key improvements:
- Live version querying: `get_stack_versions(fetch_latest=True)` now queries npm/GitHub registries to provide absolute latest versions to AI models.
- Robust version extraction: Uses a "highest version found" heuristic for GitHub Actions across all workflows.
- Scalable CLI: Passes diffs via temporary files to avoid OS command-line length limits.
- Accurate Diff Parsing: Re-implemented correlation logic in `verify_versions.py` to correctly handle multi-line version updates.
- Modular Design: Consolidated version querying and comparison logic into `utils.py` to resolve circular dependencies.
- Hard Runtime Blocks: Node.js version modifications are strictly locked unless explicitly overridden via ALLOW_NODE_VERSION_CHANGE.
- Autonomous Guardrails: AI resolutions are automatically validated and rejected if they introduce version regressions.
- Comprehensive Test Suite: New tests cover synthesized diffs, sensitive file filtering, and multi-line correlation.
